+++ src/pink/delays.clj
@@ -71,19 +71,31 @@
                (rem (unchecked-inc rw-ptr) delay-time)))
       (yield out))))
 
-;(defn frac-delay
-;  "Fixed length, interpolating delay-line. delay-time given in samples."
-;  [afn ^double delay-time]
-;  (generator
-;      [rw-ptr (long 0)]
-;      [sig afn]
-;      (let [v (aget delay-buffer rw-ptr)]
-;        (aset delay-buffer rw-ptr sig)
-;        (aset out int-indx v)
-;        (recur (unchecked-inc indx) 
-;               (rem (unchecked-inc rw-ptr) delay-time)))
-;      (yield out)))
-
+(defn frac-delay
+  "Linear interpolating (fractional) delay-line. delay-time given in samples."
+  [afn ^double delay-time]
+  (let [out (create-buffer) 
+        delay-int (long delay-time)
+        delay-frac1 (- delay-time delay-int)
+        delay-frac0 (- 1.0 delay-frac1)
+        delay-length (+ delay-int 1)
+        ^doubles delay-buffer (double-array delay-length)
+        rw-ptr (int-array 1 0)]
+    (generator
+      [rw-ptr (long 0)]
+      [sig afn]
+      (let [indx0 (let [temp-indx (- rw-ptr delay-int)]
+                    (if (< temp-indx 0) (+ temp-indx delay-length) temp-indx))
+            v0 (aget delay-buffer indx0)
+            indx1 (let [temp-indx (- indx0 1)]
+                    (if (< temp-indx 0) (+ temp-indx delay-length) temp-indx))
+            v1 (aget delay-buffer indx1)
+            v (+ (* delay-frac0 v0) (* delay-frac1 v1))]
+        (aset delay-buffer rw-ptr sig)
+        (aset out int-indx v)
+        (recur (unchecked-inc indx) 
+               (rem (unchecked-inc rw-ptr) delay-length)))
+      (yield out))))
 
 (defn adelay
   "Non-interpolating delay-line with fixed delay-time. delay-time given in
@@ -91,26 +103,12 @@
   [afn ^double delay-time]
   (samp-delay afn (int (+ 0.5 (* delay-time (double *sr*))))))
 
-;(defn fdelay
-;  "Fractional delay. Time is given in seconds."
-;  [afn ^double delay-time]
-;  (let [out ^doubles (create-buffer) 
-;        delay-buffer-len (int (+ 0.5 (* delay-time (double *sr*))))
-;        delay-buffer (double-array delay-buffer-len 0.0)
-;        read-ptr (int-array 1 *buffer-size*) ; start reading one buffer ahead
-;        write-ptr (int-array 1 0)]
-;    (fn []
-;      (when-let [asig ^doubles (afn)]
-;        (let [cur-read-ptr ^int (aget read-ptr 0)
-;              cur-write-ptr ^int (aget write-ptr 0)
-;              new-write-ptr
-;              (do-write asig delay-buffer cur-write-ptr delay-buffer-len)
-;              new-read-ptr 
-;              (do-read out delay-buffer cur-read-ptr delay-buffer-len)]
-;          (aset read-ptr 0 (int new-read-ptr))
-;          (aset write-ptr 0 (int new-write-ptr))
-;          out))))
-;  )
+
+(defn fdelay
+  "Interpolating (fractional) delay-line with fixed delay-time. delay-time
+  given in seconds."
+  [afn ^double delay-time]
+  (frac-delay afn (* delay-time (double *sr*))))
 
 
 ;; Multi-Tap Delay