+++ test/pink/delays_test.clj
@@ -6,6 +6,9 @@
             [clojure.pprint :refer [pprint]]
             ))
 
+(defn float= [^double x ^double y]
+  (<= (Math/abs (- x y)) 0.00001))
+
 (defn get-samples 
   ^doubles [afn ^long num-samples]
   (let [out ^doubles (double-array num-samples)]
@@ -25,14 +28,43 @@
   (let [afn (samp-delay (pulse 0.0) 1)
         samps (get-samples afn 200)]
     ;(pprint samps)
-    (is (= 0.0 (aget samps 0)))
-    (is (= 1.0 (aget samps 1)))
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 1.0 (aget samps 1)))
     )
   (let [afn (samp-delay (pulse 0.0) 50)
         samps (get-samples afn 200)]
     ;(pprint samps)
-    (is (= 0.0 (aget samps 0)))
-    (is (= 1.0 (aget samps 50)))
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 1.0 (aget samps 50)))
+    ))
+
+(deftest test-frac-delay
+  (let [afn (frac-delay (pulse 0.0) 1.0)
+        samps (get-samples afn 200)]
+    ;(pprint samps)
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 1.0 (aget samps 1))))
+  (let [afn (frac-delay (pulse 0.0) 1.5)
+        samps (get-samples afn 200)]
+    ;(pprint samps)
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 0.5 (aget samps 1)))
+    (is (float= 0.5 (aget samps 2)))
+    (is (float= 0.0 (aget samps 3))))
+  (let [afn (frac-delay (pulse 0.0) 1.75)
+        samps (get-samples afn 200)]
+    ;(pprint samps)
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 0.25 (aget samps 1)))
+    (is (float= 0.75 (aget samps 2)))
+    (is (float= 0.0 (aget samps 3)))
+    )
+  (let [afn (frac-delay (pulse 0.0) 5.85)
+        samps (get-samples afn 300)]
+    ;(pprint samps)
+    (is (float= 0.0 (aget samps 0)))
+    (is (float= 0.15 (aget samps 5)))
+    (is (float= 0.85 (aget samps 6)))
+    (is (float= 0.0 (aget samps 7)))
     )
   )
-