+++ CHANGELOG.md
@@ -9,12 +9,16 @@ New
 * pink.delays
   * samp-delay - non-interpolating delay line with fixed delay-time in
     samples.
+  * frac-delay,fdelay - interpolating (fractional) delay lines with fixed
+    delay-time in samples/seconds.
 
 Fixed
 
 * pink.delays
   * adelay - calculation for delay time was off by buffer-size, reimplemented
     using new samp-delay
+* pink.envelopes
+  * adsr - fixed error when \*done\* and \*duration\* were nil 
 
 ## 0.1.0